+++ Scripts/travis-build-test.sh
@@ -6,6 +6,11 @@ echo "Building"
 
 rm -rf ${HOME}/Library/Developer/Xcode/DerivedData/*
 
+if [[ "$SWIFT_BUILD" == "true" ]]; then
+    swift build
+    exit 0
+fi
+
 # -jobs -- specify the number of concurrent jobs
 # `sysctl -n hw.ncpu` -- fetch number of 'logical' cores in macOS machine
 xcodebuild -jobs `sysctl -n hw.ncpu` test -workspace Contentful.xcworkspace -scheme ${SCHEME} -sdk ${SDK} \