+++ Sources/Contentful/Date.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import ObjectMapper
 
 // Formatter and extensions pulled from: https://stackoverflow.com/a/28016692/4068264
 public extension Date {
@@ -35,19 +34,5 @@ public extension String {
     /// Return a `Date` object if the current String is in the right format.
     public var iso8601StringDate: Date? {
         return Date.Formatter.iso8601.date(from: self)
-    }
-}
-
-public final class SysISO8601DateTransform: DateFormatterTransform {
-
-    public init() {
-
-        let formatter = Date.Formatter.iso8601
-        formatter.calendar = Calendar(identifier: .iso8601)
-        formatter.locale = Foundation.Locale(identifier: "en_US_POSIX")
-        formatter.timeZone = TimeZone(secondsFromGMT: 0)
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSXXXXX"
-
-        super.init(dateFormatter: formatter)
     }
 }