+++ Sources/Contentful/Query.swift
@@ -7,8 +7,6 @@
 //
 
 import Foundation
-import Interstellar
-import CoreLocation
 
 /// The available URL parameter names for queries; used internally by the various `Contentful.Query` types.
 /// Use these static variables to avoid making typos when constructing queries. It is recommended to take
@@ -74,11 +72,26 @@ extension Date: QueryableRange {
     }
 }
 
+/**
+ Small struct to store location coordinates. This is used in preferences over CoreLocation types to avoid
+ extra linking requirements for the SDK.
+ */
+public struct Location {
+
+    public let latitude: Double
+    public let longitude: Double
+
+    public init(latitude: Double, longitude: Double) {
+        self.latitude = latitude
+        self.longitude = longitude
+    }
+}
+
 /// Use bounding boxes or bounding circles to perform queries on location-enabled content.
 /// See: <https://www.contentful.com/developers/docs/references/content-delivery-api/#/reference/search-parameters/locations-in-a-bounding-object>
 public enum Bounds {
-    case box(bottomLeft: CLLocationCoordinate2D, topRight: CLLocationCoordinate2D)
-    case circle(center: CLLocationCoordinate2D, radius: Double)
+    case box(bottomLeft: Location, topRight: Location)
+    case circle(center: Location, radius: Double)
 }
 
 /// All the possible MIME types that are supported by Contentful. \
@@ -135,7 +148,7 @@ public enum QueryOperation {
     case isGreaterThanOrEqualTo(QueryableRange)
 
     /// Proximity searches.
-    case isNear(CLLocationCoordinate2D)
+    case isNear(Location)
     case isWithin(Bounds)
 
     fileprivate var string: String {
@@ -825,7 +838,7 @@ public final class QueryOn<EntryType>: ChainableQuery where EntryType: EntryMode
                          set on the `Client` instance is used.
      */
     public convenience init<LinkType>(whereLinkAt fieldNameForLink: String, matches filterQuery: FilterQuery<LinkType>? = nil,
-                            for locale: String? = nil) where LinkType: EntryModellable {
+                            for locale: String? = nil) {
         self.init()
 
         self.parameters["fields.\(fieldNameForLink).sys.contentType.sys.id"] = LinkType.contentTypeId