+++ Tests/ContentfulTests/LocalizationTests.swift
@@ -8,20 +8,19 @@
 
 @testable import Contentful
 import XCTest
-import ObjectMapper
 import Nimble
 import DVR
 
 struct LocaleFactory {
     static func enUSDefault() -> Contentful.Locale {
-        let map = Map(mappingType: .fromJSON, JSON: ObjectMappingTests.jsonData("en-US-locale"))
-        let locale = try! Locale(map: map)
+        let usLocaleJSONData = JSONDecodingTests.jsonData("en-US-locale")
+        let locale = try! Client.jsonDecoderWithoutContext.decode(Contentful.Locale.self, from: usLocaleJSONData)
         return locale
     }
 
     static func klingonWithUSFallback() -> Contentful.Locale {
-        let map = Map(mappingType: .fromJSON, JSON: ObjectMappingTests.jsonData("tlh-locale"))
-        let locale = try! Locale(map: map)
+        let tlhLocaleJSONData = JSONDecodingTests.jsonData("tlh-locale")
+        let locale = try! Client.jsonDecoderWithoutContext.decode(Contentful.Locale.self, from: tlhLocaleJSONData)
         return locale
     }
 }
@@ -42,18 +41,21 @@ class LocalizationTests: XCTestCase {
 
 
     func testNormalizingFieldsDictionaryFormat() {
-        let singleLocaleMap = Map(mappingType: .fromJSON, JSON: ObjectMappingTests.jsonData("fields-for-default-locale"))
+        let singleLocaleJSONData = JSONDecodingTests.jsonData("fields-for-default-locale")
+        let singleLocaleJSON = try! JSONSerialization.jsonObject(with: singleLocaleJSONData, options: []) as! [String: Any]
+        let singleLocaleFields = singleLocaleJSON["fields"] as! [String: Any]
 
         let enUSLocale = LocaleFactory.enUSDefault()
-        let singleLocaleNormalizedFields = try! Localization.fieldsInMultiLocaleFormat(from: singleLocaleMap, selectedLocale: enUSLocale)
+        let singleLocaleNormalizedFields = try! Localization.fieldsInMultiLocaleFormat(from: singleLocaleFields, selectedLocale: enUSLocale, wasSelectedOnAPILevel: true)
 
         expect((singleLocaleNormalizedFields["name"]?["en-US"] as! String)).to(equal("Happy Cat"))
         expect(singleLocaleNormalizedFields["name"]?["tlh"]).to(beNil())
 
         // Multi locale format.
-        let multiLocaleMap = Map(mappingType: .fromJSON, JSON: ObjectMappingTests.jsonData("fields-in-mulit-locale-format"))
-
-        let multiLocaleNormalizedFields = try! Localization.fieldsInMultiLocaleFormat(from: multiLocaleMap, selectedLocale: enUSLocale)
+        let multiLocaleJSONData = JSONDecodingTests.jsonData("fields-in-mulit-locale-format")
+        let multiLocaleJSON = try! JSONSerialization.jsonObject(with: multiLocaleJSONData, options: []) as! [String: Any]
+        let multiLocaleFields = multiLocaleJSON["fields"] as! [String: Any]
+        let multiLocaleNormalizedFields = try! Localization.fieldsInMultiLocaleFormat(from: multiLocaleFields, selectedLocale: enUSLocale, wasSelectedOnAPILevel: false)
 
         expect((multiLocaleNormalizedFields["name"]?["en-US"] as! String)).to(equal("Happy Cat"))
         expect(multiLocaleNormalizedFields["name"]?["tlh"]).toNot(beNil())
@@ -90,3 +92,4 @@ class LocalizationTests: XCTestCase {
         waitForExpectations(timeout: 10.0, handler: nil)
     }
 }
+